+++ extras/tests/Cpp20/smoke_test.cpp
@@ -1,0 +1,15 @@
+#include <ArduinoJson.h>
+
+#include <catch.hpp>
+#include <string>
+
+TEST_CASE("C++20 smoke test") {
+  StaticJsonDocument<128> doc;
+
+  deserializeJson(doc, "{\"hello\":\"world\"}");
+  REQUIRE(doc["hello"] == "world");
+
+  std::string json;
+  serializeJson(doc, json);
+  REQUIRE(json == "{\"hello\":\"world\"}");
+}