+++ src/ArduinoJson/Json/Utf8.hpp
@@ -13,14 +13,14 @@ template <typename TStringBuilder>
 inline void encodeCodepoint(uint32_t codepoint32, TStringBuilder& str) {
   // this function was optimize for code size on AVR
 
-  // a buffer to store the string in reverse
-  char buf[5];
-  char* p = buf;
-
-  *(p++) = 0;
   if (codepoint32 < 0x80) {
-    *(p++) = char((codepoint32));
+    str.append(char(codepoint32));
   } else {
+    // a buffer to store the string in reverse
+    char buf[5];
+    char* p = buf;
+
+    *(p++) = 0;
     *(p++) = char((codepoint32 | 0x80) & 0xBF);
     uint16_t codepoint16 = uint16_t(codepoint32 >> 6);
     if (codepoint16 < 0x20) {  // 0x800
@@ -36,10 +36,10 @@ inline void encodeCodepoint(uint32_t codepoint32, TStringBuilder& str) {
         *(p++) = char(codepoint16 | 0xF0);
       }
     }
-  }
 
-  while (*(--p)) {
-    str.append(*p);
+    while (*(--p)) {
+      str.append(*p);
+    }
   }
 }
 }  // namespace Utf8