+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ArduinoJson: change log
 =======================
 
+HEAD
+----
+
+* Fixed error `'dummy' may be used uninitialized` on GCC 11
+
 v6.18.3 (2021-07-27)
 -------