+++ CHANGELOG.md
@@ -7,6 +7,7 @@ HEAD
 * Fixed error `'dummy' may be used uninitialized` on GCC 11
 * Fixed error `expected unqualified-id before 'const'` on GCC 11 (issue #1622)
 * Filter: exact match takes precedence over wildcard (issue #1628)
+* Fixed deserialization of `\u0000` (issue #1646)
 
 v6.18.3 (2021-07-27)
 -------