+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ArduinoJson: change log
 =======================
 
-HEAD
-----
+v6.18.4 (2021-09-06)
+-------
 
 * Fixed error `'dummy' may be used uninitialized` on GCC 11
 * Fixed error `expected unqualified-id before 'const'` on GCC 11 (issue #1622)